use new go-bits/audittools API
The only behavior change from this is that the debug log messages
"Notification sent" and "Notification failed" are replaced with
Prometheus counters `audittools_{successful,failed}_submissions`.

I would like to get this merged soonish to facilitate a larger cleanup
of the audittools package that turns `type AuditTrail` and `func
NewEvent` from public API into implementation details.

Note that package audittools now also offers a mock implementation that
can be used to assert on the content of generated audit events in tests.
Look for type audittools.MockAuditor for more information.